==> 1_convert.ipynb <==
import pandas as pd

#Lecture de la première feuille excel
read_file1 = pd.read_excel('data/online_retail_II.xlsx', sheet_name='Year 2009-2010')

#Lecture de la seconde feuille excel
read_file2 = pd.read_excel('data/online_retail_II.xlsx', sheet_name='Year 2010-2011')

#Concaténation des deux feuilles
read_file = pd.concat([read_file1, read_file2])

#Export du fichier final
read_file.to_csv('data/bronze.csv', index=False)
