==> 2_cleaning.ipynb <==
import pandas as pd

1. Import

df = pd.read_csv('data/bronze.csv')

2. Exploration

df.info()

df.isna().sum()

    ^ Un cinquième du dataset represente des lignes non reliées à un client. Il peut être judicieux de les conserver pour l'analyse mais on les supprimera donc avant le clustering  
    Quant au 4300 lignes sans description, on peut les conserver car la description de l'objet ne fera pas parti des features  

df.head(1)

## 3. Cleaning

    Il sera nécessaire de faire deux nettoyages.   
    - Un avant l'analyse
    - Un autre en prévision du clustering

### Analyse

3.1 Gestions des doublons

df[df.duplicated()].shape[0] # Affiche le nombre de doublons (=34335)
#Suppression des doublons
df = df.drop_duplicates()

3.2 Conversion en datetime

# L'InvoiceDate' passe de type objet à type datetime
df['InvoiceDate'] = pd.to_datetime(df['InvoiceDate'])

3.3 Customer ID

# On attribut un ID par défaut
df['Customer ID'] = df['Customer ID'].fillna(0)

###  4. Export

#Dataset nettoyé pour l'analyse
df.to_csv('data/bronze_cleaned.csv', index=False)
