==> 5_cleaning.ipynb <==
import pandas as pd

# Charger les deux fichiers CSV
silver_I = pd.read_csv('data/silver_I.csv')
silver_II = pd.read_csv('data/silver_II.csv')

# Concaténer les deux DataFrames
df = pd.concat([silver_I, silver_II], ignore_index=True)


df = df[df['Customer ID'] != 0]

#Suppression de l'écriture scientifique
pd.options.display.float_format = '{:20,.2f}'.format
df.describe()

    ^ Il y a des valeurs suspectes sur le min et le max de Quantity et de Price

Quantity

df[['Quantity', 'Description', 'Price', 'InvoiceDate']].sort_values('Quantity', ascending=False)

    ^ Les valeurs max et min en quantité sont des erreurs de commandes qui ont été annulées ensuite.

# Suppression des valeurs extrêmes pour la quantité
df = df[(df.Quantity < 20_000) & (df.Quantity > -10_000)]

Price

df[['Quantity', 'Description', 'Price']].sort_values('Price', ascending=False)

    ^ Les valeurs extrêmes de prix sont liées surcoût bancaires ou à Amazon, on va donc supprimer au-delà d'un certain palier

df = df[(df.Price > -10_000) & (df.Price < 10_000)]

StockCode

nombre_de_lignes_adjust = df[df.StockCode == 'ADJUST'].shape[0] # <- compte le nombre de lignes dont le stockcode est adjust (=67)
df[df.StockCode == 'ADJUST'].sample(2)

    ^ Ces 67 lignes ne sont pas des commandes mais des réajustements

#Suppression des lignes dont le code de stock est 'ADJUST'
df = df[df.StockCode != 'ADJUST']

Unspecified country

#Vérifications du nombre de lignes où le pays n'est pas specifié
df[df.Country == 'Unspecified'].shape[0] # = 751

#Suppression de ces lignes
df = df[df.Country != 'Unspecified']

# Diviser le DataFrame en deux parties équilibrées
df_1 = df.sample(frac=0.5, random_state=1)
df_2 = df.drop(df_1.index)

# Enregistrer les deux parties dans des fichiers CSV distincts
df_1.to_csv('data/silver_cleaned_I.csv', index=False)
df_2.to_csv('data/silver_cleaned_II.csv', index=False)
