==> 6_aggregation.ipynb <==
import pandas as pd

# Charger les deux fichiers CSV
df_1 = pd.read_csv('data/silver_cleaned_I.csv')
df_2 = pd.read_csv('data/silver_cleaned_II.csv')

# Concaténer les deux DataFrames
df = pd.concat([df_1, df_2], ignore_index=True)

aggregations = {
    'Recency' : 'first',
    'Invoice' : 'count', # = Order count -> frequency
    'Total' : 'sum', # -> Monetary
    'Quantity' : 'sum', # = Total Quantity
    'Price' : 'mean',
    'StockCode' : 'count',
    'Percent_cancelled' : 'first',
    'Average_Quantity' : 'first',
    'Order_Quantity' : 'mean', #Average_order_quantity
    'Average_Price' : 'first',
    'Average_Basket' : 'first',
    'Country' : 'first',
    'Flag' : 'first',
}

df_grouped = df.groupby('Customer ID').agg(aggregations)

df_grouped = df_grouped.rename(columns={
    'Invoice':'Frequency',
    'Total':'Monetary',
    'Average_basket':'Ø_basket',
    'Average_Quantity' : 'Ø_Qty',
    'Order_quantity' : 'Ø_InvoiceQty',
    'StockCode':'∃_items', #∃ = uniqueness
    'Price' : 'Ø_Price',
    'Cancelled_order_count' : 'n_Cancelled',
    'Percent_cancelled' : '%_Cancelled',
    })

df

# Diviser le DataFrame en deux parties équilibrées
df_1 = df.sample(frac=0.5, random_state=1)
df_2 = df.drop(df_1.index)

# Enregistrer les deux parties dans des fichiers CSV distincts
df_1.to_csv('data/gold_I.csv', index=False)
df_2.to_csv('data/gold_II.csv', index=False)
