==> 7_clustering.ipynb <==
# 1 -> K-means --> numerique
# K-proto variable mixte
# PCA -> visualisation scatter -> hue = cluster

import pandas as pd

# Charger les deux fichiers CSV
df_1 = pd.read_csv('data/gold_I.csv')
df_2 = pd.read_csv('data/gold_II.csv')

# Concaténer les deux DataFrames
df = pd.concat([df_1, df_2], ignore_index=True)

df.shape
